snc, build rules: improve behavior when compilation fails
This changes snc so that it either doesn't touch the output file at all or
else removes it if there was an error. In addition, the build rules are
changed so that output files are removed before trying to create them.

Thanks to Johnson, Andrew N. <anj@anl.gov> for analysis and patch (slightly
modified, all errors are mine).